@@ -2,10 +2,10 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>disable</Nullable>
-    <Version>8.0.2</Version>
+    <Version>8.1.0</Version>
     <Configurations>Debug;Release;Debug_Unix;Release_Unix</Configurations>
     <Platforms>AnyCPU</Platforms>
   </PropertyGroup>
